Move changlog dialog to dialog package
